separate diamond loupe from diamond
