+++ contracts/derive/lib.rs
@@ -92,3 +92,9 @@ declare_derive_storage_trait!(derive_proxy_storage, ProxyStorage, ProxyStorageFi
 // Diamond
 #[cfg(feature = "diamond")]
 declare_derive_storage_trait!(derive_diamond_storage, DiamondStorage, DiamondStorageField);
+#[cfg(feature = "diamond")]
+declare_derive_storage_trait!(
+    derive_diamond_loupe_storage,
+    DiamondLoupeStorage,
+    DiamondLoupeStorageField
+);
